Changes necessary for consistent data product handling.
We attempt to at least partially deal with two interlocking problems:

1. Since QuickVec.hh is compiled as a template, the return value of
@Class_Version()@ _must_ be updated every time private member data are
changed. If it is compiled as a non-template, then @classes_def.xml@
would need to be changed to add the ClassVersion attribute, which would
thenceforward be managed by the build system's @checkClassVersion@
mechanism. If you no longer intend to use the option of the non-template
compile of QuickVec.hh, I would suggest removing the conditional
compilation in order to simplify the code.

2. Between ROOT 5.34.18 and 5.34.21, (s3, s4+), ROOT's checksum system
changed to incorporate base classes of data products, which means that
the checksum will be different for some products in s3 or lower than for
s4 or higher. I have attempted to account for this with class versioning
(see the checksums for versions 13, 14 and 15 of @Fragment@), but the
safest course will be to withdraw support for s3 builds of future
versions of artdaq_core.