Currently-deactivated plugin for artmod successor.
This directory contains a technology preview of a skeleton generator
plugin for the new cetlib "cetskel" utility to produce a skeleton for
the artdaq-core "generator".

The subdirectory is not yet added to the top level CMakeLists.txt file,
so this addition is zero impact. When activated, however, it will
require modern versions of cetlib and upwards.